requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
base58==2.1.1
bitarray==2.6.0
bitcoinlib==0.6.5
branca==0.5.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
cytoolz==0.12.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.2.1
django-crispy-forms==1.14.0
django-fsm==2.8.1
django-getpaid==2.3.0
django-getpaid-payu==0.2.1
django-model-utils==4.2.0
django-paypal==2.0
django-social-share==2.2.1
ecdsa==0.18.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
folium==0.13.0
frozenlist==1.3.1
greenlet==1.1.3.post0
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jsonschema==4.16.0
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.23.4
opencage==2.1.0
parsimonious==0.8.1
pendulum==2.1.2
phonenumbers==8.12.57
Pillow==9.2.0
protobuf==3.19.5
psycopg2-binary==2.8.6
pycparser==2.21
pycryptodome==3.15.0
pyOpenSSL==22.1.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2019.3
pytzdata==2020.1
#pywin32==304
requests==2.28.1
rlp==2.0.1
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.3.0
swapper==1.3.0
toolz==0.12.0
typing-extensions==3.10.0.2
tzdata==2022.5
urllib3==1.26.12
varint==1.0.2
web3==5.31.1
websockets==9.1
whitenoise==5.2.0
yarl==1.8.1
